# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ai-agent/memories": {
      "post": {
        "operationId": "AiRecordsController_createMemory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "value": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "sessionId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create own memory",
        "tags": [
          "AI Records"
        ]
      },
      "get": {
        "operationId": "AiRecordsController_listMemory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Own user ID only; tenant admins may filter audit logs by another member",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped paginated records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "page",
                    "pageSize"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "tenantId": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "key": {
                            "type": "string"
                          },
                          "value": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "sessionId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "version": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paginate own memories; optional userId filter cannot widen scope",
        "tags": [
          "AI Records"
        ]
      }
    },
    "/api/ai-agent/memories/{id}": {
      "get": {
        "operationId": "AiRecordsController_getMemory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "value": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "sessionId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get own memory by UUID",
        "tags": [
          "AI Records"
        ]
      },
      "patch": {
        "operationId": "AiRecordsController_updateMemory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "value": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "sessionId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update own memory",
        "tags": [
          "AI Records"
        ]
      },
      "delete": {
        "operationId": "AiRecordsController_deleteMemory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete own memory",
        "tags": [
          "AI Records"
        ]
      }
    },
    "/api/ai-agent/tasks": {
      "post": {
        "operationId": "AiRecordsController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "sessionId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "type": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "running",
                        "completed",
                        "failed",
                        "cancelled"
                      ]
                    },
                    "progress": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    },
                    "input": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": true,
                      "nullable": true
                    },
                    "errorSummary": {
                      "type": "string",
                      "nullable": true
                    },
                    "version": {
                      "type": "integer"
                    },
                    "startedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "finishedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create own task",
        "tags": [
          "AI Records"
        ]
      },
      "get": {
        "operationId": "AiRecordsController_listTask",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Own user ID only; tenant admins may filter audit logs by another member",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped paginated records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "page",
                    "pageSize"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "tenantId": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "sessionId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "type": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "running",
                              "completed",
                              "failed",
                              "cancelled"
                            ]
                          },
                          "progress": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 100
                          },
                          "input": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "result": {
                            "type": "object",
                            "additionalProperties": true,
                            "nullable": true
                          },
                          "errorSummary": {
                            "type": "string",
                            "nullable": true
                          },
                          "version": {
                            "type": "integer"
                          },
                          "startedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "finishedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, pagination, or attempted ownership override"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Tenant membership or administrator permission required"
          },
          "404": {
            "description": "Record not found within the authorized scope"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paginate own tasks; optional userId filter cannot widen scope",
        "tags": [
          "AI Records"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dashboard API",
    "description": "Authentication, business management, dashboard aggregation, tenant-isolated AI records, conversations and read-only AI tools.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateMemoryDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "value": {
            "type": "object",
            "additionalProperties": true
          },
          "sessionId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "key",
          "value"
        ]
      },
      "UpdateMemoryDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "value": {
            "type": "object",
            "additionalProperties": true
          },
          "sessionId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "input": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "type"
        ]
      },
...
    }
  }
}